build fix.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@53489 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/socket.cpp
@@ -1618,7 +1618,7 @@ int rtSocketBindRawAddr(RTSOCKET hSocket, void const *pvAddr, size_t cbAddr)
     AssertReturn(rtSocketTryLock(pThis), VERR_CONCURRENT_ACCESS);
 
     int rc;
-    if (bind(pThis->hNative, (struct sockaddr const *)pvAddr, cbAddr) == 0)
+    if (bind(pThis->hNative, (struct sockaddr const *)pvAddr, (int)cbAddr) == 0)
         rc = VINF_SUCCESS;
     else
         rc = rtSocketError();
